backend/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "统一API网关系统"
authors = [
    {name = "API Gateway Team"}
]
dependencies = [
    "fastapi[all]>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.1",
    "websockets>=12.0",
    "structlog>=23.2.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "cryptography>=41.0.7",
    "httpx>=0.25.2",
    "python-multipart>=0.0.6",
    "paho-mqtt>=2.1.0",
    "alembic>=1.16.5",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "psutil>=5.9.8",
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["docker*", "tests*"]

[tool.setuptools.package-data]
app = ["**/*.sql", "**/*.yaml", "**/*.yml", "**/*.json"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=app"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default= true
